// Angular/EmployeeApp/src/app/models/employee.model.ts
export class Employee {


    empno: number;
    ename: string;
    job: string;
    mgr: number;
    hiredate: Date;
    sal: number;
    comm: number;
    deptno: number;


    /* id: number;
    name: string;
    gender: string;
    email? : string;
    phoneNumber?: number;
    contactPreference: string;
    dateOfBirth: Date;
    department: string;
    isActive: boolean;
    photoPath?: string; */
}